Adjust behavior if change to different mode.
